# Improve performance of `setproperties`/`getproperties` for structs with unions