Addresses #55.

These changes make use of generated functions in the case where properties are fields, both for `getproperties` and `setproperties`.

In terms of performance, see
```julia-repl
julia> using ConstructionBase

julia> using BenchmarkTools

julia> struct A
           a::String
       end

julia> x = A("hello");

julia> @btime getproperties($x)
  1.259 ns (0 allocations: 0 bytes)
(a = "hello",)

julia> @btime setproperties($x, (; a = $("world")))
  1.433 ns (0 allocations: 0 bytes)
A("world")

julia> struct B
           a::Union{Nothing, String}
       end

julia> x = B("hello");

julia> @btime getproperties($x)
  53.251 ns (2 allocations: 32 bytes)
(a = "hello",)

julia> @btime setproperties($x, (; a = nothing))
  0.759 ns (0 allocations: 0 bytes)
B(nothing)

julia> @btime B(nothing)
  0.719 ns (0 allocations: 0 bytes)
B(nothing)
```

`getproperties` still allocates, this is due to using a `NamedTuple` that contains a union for one of its fields. I guess one could force the `Union` to be formed on the outside, e.g. `Union{NamedTuple{....}, NamedTuple{...}}`, but I think the current state is fine and has the benefit of simplicity. `setproperties` is as fast as one would expect, at least.

The `"no allocs S2"` testset captures the performance improvements of this PR, failing on current `master`.