@serenity4 2024-07-28
The [Setfield failure][https://github.com/JuliaObjects/ConstructionBase.jl/actions/runs/10131761234/job/28014870173?pr=91#step:6:297) is due to some timing comparison. The test passes locally for me on the same Julia version, with much lower timings (7ns and 1ns), so I'm guessing it might just be some runner performance inconsistency.