@rafaqz 2024-07-29
Does `getproperties` on a different Union have that cost in a larger context where the `NamedTuple` is never returned from a function?

It probably won't need that allocation in practice.