@serenity4 2024-08-13
Any objections to merging this? AFAIK this should no longer break anything and only result in performance improvements, particularly so for `setproperties`.

If the small changes to `getproperties` aren't welcome, I'm fine with removing them, the intended optimization target was mainly `setproperties`.